refactor(cannon): move execution context into cannon/context.rs
